LCD: Add project
The TFT-LCD screen is from ALIENTEK(正点原子).
I copied the driver (lcd.c/lcd.h/font.h) from samples and do some migrations.
The FSMC uses NE1, so the address is from 0x60000000 to 0x63FFFFFF.
The connector on my board uses register A18(to PD13),
so the base address start from 0x7FFFE
(contains 18 '1' and a trailing '0', based on the given algorithm).
The FSMC initialization is done by the code of CubeMX, so I remove the relative part in lcd.c/.h.